Lots of Fixes, Added QR codes and jQuery
Did a lot of repair of code and cleaning, fixed up the presentation
with jQuery UI.  added some jQuery for the Preview of the label print
out.  now the page can be printed out with out the form showing.
Moving to the LGPL license 3, the same as the PHPQRcode.  Updated the
README file.